frontend/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-bigquery-storage
streamlit
db-dtypes
matplotlib
plotly
google-genai
streamlit-card
langchain
langchain-google-vertexai
langchain-google-community
langgraph
python-dotenv